refactor(storage): extract raw revision/membership governance from archive.py
Problem: archive_tiers/archive.py (13,338 lines) mixes its documented
query-surface contract with ~55 raw-revision-authority and membership-
classification methods -- a different concern with different invariants.
Every 2026-07-30 defect (PRs #3394/#3396/#3397/#3398/#3401) landed in this
cluster or its callers (polylogue-c737 et al, polylogue-1r9c hotspot map).

Solution: moved the raw revision/membership governance surface to a new
module polylogue/storage/revision_governance.py as free functions taking
`store: RawRevisionGovernanceHost` (a Protocol naming exactly the seven
ArchiveStore members governance code touches: _conn, _ensure_source_conn,
_blob_publisher, _pending_raw_parse_states, _preacquire_attachment_blobs,
_write_counts, _skipped_counts) instead of `self`. ArchiveStore keeps
one-line delegating methods with unchanged signatures, so every external
caller (sources/live/batch.py, sources/live/append_ingest.py,
sources/revision_backfill.py, storage/repair.py,
pipeline/services/archive_ingest.py, api/archive.py, and every test holding
an ArchiveStore instance) is untouched.

Considered and rejected: (a) passing the bare sqlite3.Connection --
insufficient, governance needs the lazy source-conn, blob publisher, and
pending-raw-parse-state batch too; (b) a mixin ArchiveStore inherits from --
rejected because inheritance gives every moved method unrestricted `self`
access to the other ~9,000 lines of read-surface internals, exactly the
"reach back into internals" the extraction is meant to make impossible.

archive.py: 13,338 -> 11,318 lines (-15.1%). New module: 2,838 lines
(documented contract + moved implementation).

Verification: mypy --strict clean on both files; ruff format/check clean;
python -c "import polylogue.storage.sqlite.archive_tiers.archive" succeeds.
